backend/authservice/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "passlib[bcrypt] (==1.7.4)",
    "bcrypt (<4.0)",
    "aio-pika (>=9.6.2,<10.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
